+++ .vscode/extensions.json
@@ -3,7 +3,8 @@
 		"dbaeumer.vscode-eslint",
 		"unifiedjs.vscode-mdx",
 		"graphql.vscode-graphql-syntax",
-		"editorconfig.editorconfig"
+		"editorconfig.editorconfig",
+		"vitest.explorer"
 	],
 	"unwantedRecommendations": []
 }
