+++ scripts/package.json
@@ -1,13 +1,15 @@
 {
 	"name": "@swaglgbt/swaglgbt-scripts",
+	"description": "Various scripts to help with common tasks and CI",
 	"private": true,
 	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1"
+		"test": "vitest"
 	},
 	"author": "Cass Fridkin <cass@swag.lgbt>",
 	"license": "PPL",
 	"devDependencies": {
 		"@types/github-script": "github:actions/github-script",
-		"@types/node": "^20.12.11"
+		"@types/node": "^20.12.11",
+		"vitest": "^1.6.0"
 	}
 }
