+++ scripts/tsconfig.json
@@ -1,4 +1,4 @@
 {
 	"extends": "../tsconfig.json",
-	"include": ["**.mjs"]
+	"include": ["**/**.mjs"]
 }
